Comments out migration with syntax error in it
File 202008211813.do.nys-privacy-v3.sql contains a syntax error that causes DB migrations to fail. Due to how the NPM migrations module works I'm not 100% sure that removing the migration file will result in a valid migration. Instead I've commented out the UPDATE statement and have replaced it with a SELECT of no data to ensure something is run in the file that will not cause an error.